Modificacion pasando estado de los tweets a un componente padre para Tweets y para crear Tweets
