+++ App-React-Twitter/src/components/Tweets.jsx
@@ -1,20 +1,9 @@
 import CardTweet from "./CardTweet"
-import React,{ useState, useEffect } from "react";
-import CrearTweet from "./CrearTweet";
+import React from "react";
 
 
-function Tweets(){
-    const [tweets, setTweets] = useState(null)
+function Tweets(props){
 
-    useEffect(() => {
-        getTweets();
-    }, []);
-
-    const getTweets = async() => {
-        const dataExistente= await localStorage.getItem('tweets')
-        let parseados=JSON.parse(dataExistente)
-        setTweets(parseados.reverse())
-    };
 
 
     return(
@@ -28,12 +17,11 @@ function Tweets(){
                 <h1 className="m-5 font-bold">Tweets</h1>
 
                 <section className="bg-azulOscuro2 rounded-t-lg w-full h-full">
-                    {tweets && tweets.map((data)=>(
-                        <CardTweet key={data.id} data={data} actualizarTweets={getTweets}/>
+                    {props.tweets && props.tweets.map((data)=>(
+                        <CardTweet key={data.id} data={data} actualizarTweets={props.actualizarTweets}/>
                     ))}
                 </section>
             </section>
-            <CrearTweet actualizarTweets={getTweets}/>
         </div>
     )
 }
